requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.0
transformers==4.35.2
torch==2.1.1
torchaudio==2.1.1
torchvision==0.16.1
datasets==2.14.6
accelerate==0.24.1
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3
pandas==2.1.3
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
elasticsearch==8.11.0
redis==5.0.1
celery==5.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.0.3
httpx==0.25.2
websockets==12.0
ffmpeg-python==0.2.0
moviepy==1.0.3